Fix the issue identified by Pokuri regarding crashes when multiple threads are running.  The Pmml object that was holding a reference to the Context object needed to update data dictionary and transaction dictionary values via function call was being hammered as multiple threads were updating the reference.
To fix, the Context object has been added to the parameter list of these functions.  We do not directly expose
this runtime context object in the Pmml, however.  Instead, the 'ctx' is added to the parameter list via
a macro.  The macro 'Put(variableName, someValue)' causes 'Put(ctx, variableName, someValue)' to be generated
in the scala output.