New version of the copd model - COPDv1.xml - relies principally on icd9 codes.  CodeSets used for filtering various conditions made available through EnvContext.  This model compiles cleanly.
Dyspnoea.csv now one of the filter tables (inadvertently omitted in earlier checkin).

New udfs introduced to support the filtering (Contains, IsIn, ContainsAny, AsCompressedDate). Some screwups on
the bootstrap and cmstypes corrected.  A bug detected and fixed in the FunctionSelect. In certain circumstances,
the translation between a PMML standard builtin function name (e.g., *, -, +) was not being translated to the
pmml udf library implementation for it. Added Dyspnoea type to the list of supported types in the provisional
KVInit we currently have.

Brief note regarding the use of debugging directory, modeldbg also added.