Implement simple array extract of the container values for the supplied map name (impl for getArrays). Change the "container" name in the model from "filterMap" to "filterArray" to more properly reflect what is returned.
Note: The "filterArray" container key is not used.  The key is really what is used to pull the values
for the entire map (i.e., the key is the name of the container).  This is something of a hack doing this
way.  We need to revisit EnvContext protocol to allow a more straight forward usage where perhaps only
one argument is supplied for getArray for example in order to get all of the BaseContainer values.

For the pending release, we change the getArray around a bit to deal with this.