Pmml Model, PmmlCompiler, PmmlUdfs, MedEnvContext, Metadata project changes to support models that have "look back" history processing.
1. Fixed function AsCompressedDate - month shift off by factor of 10 .
2. COPD model .  Not possible to downcast directly from Array[BaseContainer] to Array[SmokeCodes] (as an example) Changed
model to use DownCastArrayMembers macro that cast the individual array items with a map function.
3. Added new <Constant> dataTypes... mbrTypename and typename.  What these print, they don't print the <Constant>'s value but
use it instead as a field name.  The field name is looked up and its member type or type is printed respectively.  Only fields with
Array and ArrayBuffer are supported for mbrTypename.  This feature is needed for the DownCastArrayMembers macro.
4. New Metadata typedef added for immutable sets.  The ContainerMap function for the derived field SputumCodeSet (for example) produces
an immutable set.  It can't be cast to the mutable sets we had been using.  Hence immutable version of set added.
5. YearsAgo(numYrs : Int) function added returns millisecs since epoch.
6. MedEnvContext has cache of Array[BaseContainer] for each of the COPD filter types (Smoke, Cough, et al)
7. I have checked in the .csv version of William's messages_new_format_all.csv.gz

	WARNING: William, I didn't change the file name.. so when the gz is produced by the install script, your
	notes will be wiped out.  Preserve the expected results in some other way I think.. maybe we can just add additional
	field to the csv lines themselves to indicate what each row would produce.

BTW, only the 1a result is produced by processing messages_new_format_all.csv.gz with these changes.  Is there a logic
error in model?  Or is the data values not sufficient to produce 1b?  I check this in so more eyes can look and help isolate.