Changed name of type to match the EnvCode db.  This was a problem that would cause problems with lookup of EnvCode values.
HOWEVER, no difference in results.  I verified that each of the tables were being loaded by stopping in the generated model functions at both the extract and set creation.  That looks fine.   If it is code issue, I would suspect that there is something wrong with the udfs for 'IsIn' or 'ContainsAny'... I will stop in these functions to see if that is the case next.

I also verified that beneficiary records were being processed correctly in this way... I stopped in the Age function to see if the age changed during procesing.  I processed a one that is 70 yrs and one that is 49 yrs.

The other possibility is to that msgs are being excluded due to the 1 year lookback period (clm_thru_dt is too old).