Add new option to arg list for PmmlCompiler... --instrumentWithLogInfo <value>
If you put a 't' or '1' or 'y' as the value of --instrumentWithLogInfo, it will generate a model that
has an "enter" message into each derived field's execute and a message that prints the result.

If that option is left unspecified, no logging messages are injected into the model.  The log level
used is "info" btw.

I have fixed the call to the pmml compiler in CompilerProxy and TestMetadataAPI... simply by setting
the value to PmmlCompiler ctor to false.  This has been tested and messages come out in the OnLEPManager log.