Several bug fixes in the FunctionSelect type inference.  I also implemented the env variables in MetadataAPI.
I did this to keep William's test data in same place and also not pollute the MetadataAPI examples with
my experiments.  Variables added are :

		MODEL_FILES_DIR"   		: "/home/rich/william",
		"TYPE_FILES_DIR"      		: "/home/rich/william",
		"FUNCTION_FILES_DIR"      	: "/home/rich/william",
		"CONCEPT_FILES_DIR"      	: "/home/rich/william",
		"MESSAGE_FILES_DIR"      	: "/home/rich/william/messages",
		"CONTAINER_FILES_DIR"      	: "/home/rich/william/containers"

For the FunctionSelect bugs, I honestly don't know how one of these problems did not present itself earlier.
The first superclass in the list of superclasses was always taken, but that is always the same as the original
class that was used to fetch the superclasses from the scala runtime mirror...

There remains a problem that is caused by not having an ImmutableMapOfIntArrayOfInpatientClaim type.  Ramana
is going to add it shortly.  This fix will be used in next commit coming shortly.