MethodExtractor improvements:
1) MethodExtractor was producing all of the types, not the unique set of types needed by the methods in the UDF object.
2) MethodExtractor now can produce a fat jar, which is used by the...
3) New script, extractUdfLibMetadata.scala,  that packages the sbtProjDependency.scala usage inside it to avoid the user having to use the error prone syntax directly

These two scala scripts should be placed on your PATH.  They are executed with bash (no windows version).

One more round on this is needed to generalize the path usage.  The plan is to make the paths relative to the current working directory.  One must be
in the top level sbt project folder (e.g., ~/sbt/RTD/trunk) in order for this script to run properly.

Sample syntax: