PmmlCompiler Cleanup and generalization of the udf metadata extraction
Specifically:

1) UdfExtract : sort the required types that were gleaned from parse of the supplied UDF object
2) PmmlCompiler : eliminate the need for the jar manifest file.. since we are not using it in the loading process.  Make sure that no
bad return codes are missed from the production of the jar.
3) extractUdfLibMetadata.scala : improve the help documentation. Use the present working directory to express the udf extract
application location and other parameters needed to create the UDF metadata.
4) installOnLEP_Medical.bat and installOnLEP_Medical.sh - Add the creation of the MethodExtractor fat jar to the scripts.  Since
this jar is dev side, it is not copied to the OnLEPManager's install location.

MethodExtractor usage:

To create the Udf metadata for the core udf lib in trunk/Pmml/PmmlUdfs, the following command is used:

	extractUdfLibMetadata.scala --sbtProject PmmlUdfs --fullObjectPath com.ligadata.pmml.udfs.Udfs --namespace pmml --versionNumber 100 > /tmp/json.txt

where
	extractUdfLibMetadata.scala (and the sbtProjDependencies.scala) script must be on your PATH.  The sbtProject
	name is mentioned and the full package qualifed object name where the udfs reside are mentioned.  You can
	choose the namespace and version number to use in the json output for each type and
	function definition.

All output is currently coming out to stdout with the types first followed by the function def jsons.  There is a blank line between the types and the udf json strings.
Currently there is no script tailored for Windows.