failed to put the version suffix on the class and object names for generated models.
This has been now tested with the new data messages50_2014_BIOH.csv which produces 28 positive results.
It should be noted that this only currently works with "hashmap".  When your configs have "treemap" the
EnvContext implementation does not load the .db files instead creating EMPTY .hdb files.  This has the
unpleasant effect of not producing any output (indeed the icd9codes are not found in those empty .hdb stores).

We need a little more sophistication for this.  The KVInit app should be able to produce either hdb, db,
or cassandra (ultimately).  The SimpleEnvContext needs to be able to recognize what to do when
confronted with the task of loading containers.