Cleaned up version of the banking model... model runs but produces no alerts with files that are known to produce data formerly.
Take a look at the README.md for how this is organized.

The Alchemy model and the kv store containers are loading... stopped in debugger to verify.
There is a problem with the data. I am not sure why, but only the first record in the
msg file (I was using OneCustomerTransactionsAlerts_LBOD123EB2.csv which should produce
a LB, an OD1, an OD2, an OD3 and an EB2 alert) is the only record seen by the model.
Unfortunately, the first record produces no alerts.

This seems crazy, but to prove this to myself, I added the model to the project so I could
put breakpoints at the LB test (the first alert is an LB produced by 2nd record in file).
The run ledger balance never changes from 110.0 which is the balance on the first record.

I will try to figure it out, going back into the engine and seeing why the next record is
not being processed.

The config I am using for OnLEPManager is BankPOC_LBOD123EB2.cfg which goes with the
OneCustomerTransactionsAlerts_LBOD123EB2.csv data file.