Error location information is now emitted with semantic errors.
A stack of the current syntax tree node path is being maintained during compilation.  Should
an error be detected, this information is emitted as a prologue to the error encountered.  The
messages in the log take the form "While processing Apply function 'foo' found in Apply function 'bar'
found in DerivedField 'doh'..."