The load of the the functions from MethodExtractor generated typedef and fcndef json files now works pretty well... good enough for med client to add their functions.  For the core udf lib we have, 693 functions out of roughly 700 functions are being cataloged. There is an issue in json generation in the method extractor for immutable maps that, when fixed, will allow us to delete nearly all of the function definitions in the MetadataBootstrap and simply load them from the MethodExtractor generated json (i.e., totally eating our own dog food at that point).  From the perspective of the medical customer, the current functionality should be sufficient to meet their needs.  They only need to create their messages, containers, and concepts and make the generated jars part of their UDF sbt project (perhaps putting them in the lib folder as "unmanaged jars")
Specific changes:

MetadataAPI - Changed the function signature for GetAllFunctionDefs
-  def GetAllFunctionDefs(formatType: String) : String
+  def GetAllFunctionDefs(formatType: String) : (Int,String)

JsonSerializer... the function def array was changed to an array buffer so the functiondefs could be collected... prior way used
immutable container that never collected the function defs.

MetadataLoader... com.ligadata.Pmml.Runtime.Context added .  The bootstrap of the functions are still in place here.  There is an
issue with generating immutable arg version of MapValues and MapKeys (bug in MethodExtractor that generates the json for the function defs).

TestMetadataAPI - printing the number of functions in metadata on "dump all functions" option.

MetadataAPIImpl - returning the number of types and functions cataloged from the "load from file" options.  In the case of the
types, nearly all of the types for the core udf lib are already declared in the bootstrap.  Only four are actually cataloged; the
others are noted as "already there" Both the load type and load fcn from file now works due to the fixes (use of ArrayBuffer instead
of immutable.Array) in the JsonSerializer.

PmmlUdfs - the MultiSet functions are commented out for now.  They were not cataloging properly due to the complex type arguments.
I suspect this is either the same or another bug in the MethodExtractor.

coreUdfFcnDefs.json and coreUdfTypeDefs.json - the current version of these is checked in.  These are generated files by the
methodExtractor script:

   extractUdfLibMetadata.scala  --sbtProject PmmlUdfs
				--fullObjectPath com.ligadata.pmml.udfs.Udfs
				--namespace pmml --versionNumber 100
				--typeDefsPath ~/github/Med/RTD/trunk/SampleApplication/Medical/Types/coreUdfTypeDefs.json
				--fcnDefsPath ~/github/Med/RTD/trunk/SampleApplication/Medical/Functions/coreUdfFcnDefs.json