Cluster installation script and its constituent parts.
clusterInstallOnLEP.sh - top level script
NodeInfoExtract.scala - extract node configuration info for the which machines and which nodes and which directories get installations
nodeInfoExtract.sh - helper script that organizes call to the underlying scala... prepares the current classpath with sbtProjDependencies script.

The clusterInstallOnLEP.sh script should be run from the trunk directory of the repo containing the release.  It should be run from the account
that will ultimately be used to manage the OnLEP cluster.  All machines in the cluster must be preconfigured to run ssh and scp commands
in a password-less fashion.