Driving Digital drop in for easy install... model tested and produces output.  Data had to be manufactured to get logic to work... this aspect still needs some work... A new variable needs to be added for the message sent ... it needs to be supplementary not prediction... will fix shortly
Changes:

1) DrivingDigitalApp directory... modeled like the input/application1 sample and intended to be copied into the <installdir>/input/
This is not hooked up in any way with the string substitution for the easy install.  The directories in cfg are set to /tmp/drdigital.
There are notes in DrivingDigitalApp/builddrdigital.notes.txt.

2) Bug fixes for some date udfs ... copy / paste errors plus bad DateTime ctor use.

Note: The old Alchemy and original (and BAD) DrivingDigital still exist in repo at trunk/SampleApplication/Bank... don't use.. Use what
is in trunk/SampleApplication/Bank/DrivingDigitalApp