Several ERROR log messages improved by prefacing the error msg with error location stack information.
PmmlError.logError(...) replaced log.error(...) which prints the derived field and apply hierarchy where the error was
encountered.  It is desirable to print the line number where the error occurs, and that may be added later.  I have
to see if line number is available from the SAX parser for each element it sees.  If that is the case, then error numbers
will be trivial to add.