Fix for Issue #138 ("If() function generates incorrect code") and at least part if not all of Issue #134 ("If() function not Pmml-4.1 compliant") and Issue #127 (#127 If() function not properly generated)
1) Changed If function signature to template as Pradeep suggested in Issue #134
2) Partial fix for use of camelcase and pascalcase names in the models... more work to do ... see Issue #141
3) Semantic check on if elements ... there must be 3 children (predicate, true action and false action)

Notes:
1) Only "if (pred) trueaction else falseaction" forms will be supported for now. The procedural form,
"if (predicate) trueaction" will be not be supported at least until more thought is given as to how
to handle the false case.  As was pointed out by Pradeep, the action results - both the true and false ones -
can be any type.  The question then is what would the default value be?  The PMML, like Scala,
is a functional language wanting to return a function result of some type.

2) The If function can return any type.  Sample tests have been embedded in the model that revealed the
error.  It will be posted to the Issue 138 thread for your inspection. There are 4-5 test cases just
before the monster derivedfield that has been troublesome.

3) The model now builds after a few name changes, use of 'or' and 'and' in lieu of the contains{any,all}
functions that don't exist, and numerous mis-specifications in the monster all related to the use
of And(boolexpr).  And is a binary operation (we don't have And(Boolean))... only And(Boolean,Boolean)
and others up to 7-8 arguments.