Fix for #162, #138 'if' pmml elements now generates native scala "if <pred> { <tAction> } else { <fAction> } form.
Notes:
1) special handling for if in node printers
2) If udfs removed form PmmlUdfs and bootstrap
3) or(Int*):Int and and(Int*):Int changed to IntOr(Int*):Boolean and IntAnd(Int*):Boolean respectively
4) COPDv1.xml that used or(Int*) changed to use IntOr(Int*)

Important:
The if pmml elements still expect three arguments.  We are not yet supporting the 2 argument form (if <pred> <tAction)
For the IndefiniteArity change, only a few of the udfs (and, or, intand, intor) have been changed.  Another pass through
the udf lib and the bootstrap will be made to convert other indefinite arity function candidates... Plus, Multiply, et al