Provisional fix for #162 (previously #160 and #161).  The "or" and "and" variadic functions now generate inline scala "||" and "&&" respectively.
This is a general fix as discussed in #162 thread, but provides short circuit for two of the most common use cases seen in the dsl generated
code.  This will allow most of the nested "if" to be converted to (if (expr && expr && expr && ...)" or "||" ops as the case may be.