Fix for #195.  Add macro GetXid to make transaction id from engine directly available in the model.
In your Pmml.xml add this derived field:

	<DerivedField name="CurrentXid" dataType="Long" optype="continuous">
		<Apply function="GetXid"/>
	</DerivedField>

You may then reference it from your function call with a FieldRef.

Note: There is no builtin type in PMML for 'long'.  As such, putting the macro reference
inline in your function call you are building mentioned in #195 will not work.  By putting
it in a DerivedField, however, the generated code that will access it in the
new function will generate the correct coercion.