A UDF "search path" is introduced.  In the PMML model one can add additional custom udf libraries by adding their full package qualifed object name to a variable in the Pmml model's DataDictionary. This variable must be named (exactly) UDFSearchPath.  For example,
    <DataField name="UDFSearchPath" displayName="UDFSearchPath" dataType="container">
      <Value value="com.ligadata.pmml.udfs.EdifecsPmmlUdfs" property="valid"/>
    </DataField>

added to the DataDictionary section will cause EdifecsPmmlUdfs to be imported in the generated model,
thereby making the UDFs available.  Note that there is no need to specify the core udf library
here.  This is strictly for other Scala Objects that contain custom UDFs needed in the model.

More than one library can be made available in this way.  Simply add a second Value line with a different
library mentioned.  Currently only Scala UDF objects have been tested. The intention is to support
java static methods as well. Experimentation is under way.

Note that no attempt is currently made to scrub duplicates.  We will add that.  For now, a duplicate
import will likely produce a warning by the Scala compiler.