Fatafat/Fatafat-2 task - Install cluster, start cluster, stop cluster, status cluster, log error scanners (2 of them)
This commit satisfies the Jira task Fatafat-2 for this sprint.  Pokuri has vetted the installer, start, stop,
and status scripts.  In addition two log scanner tools that extract ERROR messages from the logs of all of the
FataFat cluster nodes for a given cluster have been developed and tested by me.  The log errors for each node
are aggregated on the admin machine for ALL of the nodes in the cluster.  The elementary version will scan
an entire log.  The other will accept a time unit (e.g., 5 minutes) and only look at records between 5 minutes
ago and right now.

The documentation found in README.md has been updated to reflect the changes to the installer script we made
as it was vetted on the customer machines by Pokuri. Additional documentation for each of the other scripts has now
been added and can be incorporated in the wiki.

A small tool was developed to aid in generating a lookbackTimeStamp from which to limit the log scan.  See DateCalc.