Macro code produced mapped instead of fixed code generation
Ahmed discovered an issue with certain macro usage in the pmml.  During debug, I discovered that
the wrong code was being generated for the offending derived field.  Mapped code was being incorrectly
generated.  This has been fixed and is now available in "tweetie" - a branch off of dev.

This should be used either directly for the twitter demo or its changes merged to whatever will be used.