Macro fix of fixes - fixed messages use fixed template, mapped use mapped template
First attempt overcompensated s.t. for Put, the fixed message is always used (EnvContext is a fixed container).  This
has been changed.  If there is a mapped message in the macro now, the mapped template  will be chosen.

Several of the mapped macros did not coerce the Any type to its rightful type... asInstanceOf coercion
added for these.

NOTE:
We have one macro, setField, that accepts two containers with de-references.  Currently BOTH of the containers must either
be both dereferences of FIXED messages or containers or both must be MAPPED.  This is not ideal.  The medical client is going
all mapped afaik, so we should not see an issue with their use.  Most applications will use FIXED.

We need to schedule a refurb on the macros to recognize the dereference types individually and possibly generate the appropriate
combination of mapped and fixed dereferences.