Fix #395. classes with same name in same scope
A map has been added to watch what classes are generated for the Builds/Does macros.  The names
used in the macro template now use all fields for the class name to disambiguate names that
update the same field with three and four argument macros.

I tested with a modified version of the twitter_demo pmml that updated the same fields many times.
I will send this to the twitter team and to William and anyone else who might wish to see.