While joda's DateTime object is not explicitly available as a type in the pmml, these functions are now available:
	def dateDaysSinceYear(yr: Int): Long = {
	def dateSecondsSinceYear(yr: Int): Long = {
	def dateSecondsSinceMidnight(): Long = {
	def dateMilliSecondsSinceMidnight(): Long = {
	def Timenow(): Long = {
	def AsSeconds(milliSecs: Long): Long = {
	def Now(): Long = {
	def YearsAgo(numYrs: Int): Long = {
	def YearsAgo(someDate: Int, numYrs: Int): Long = {
	def YearsAgoAsISO8601(someDate: Int, numYrs: Int): Int = {
	def MonthsAgo(numMos: Int): Long = {
	def MonthsAgo(someDate: Int, numMos: Int): Long = {
	def MonthsAgoAsISO8601(someDate: Int, numMos: Int): Int = {
	def WeeksAgo(numWks: Int): Long = {
	def WeeksAgo(someDate: Int, numWks: Int): Long = {
	def WeeksAgoAsISO8601(someDate: Int, numWks: Int): Int = {
	def DaysAgo(numDays: Int): Long = {
	def DaysAgo(someDate: Int, numDays: Int): Long = {
	def DaysAgoAsISO8601(someDate: Int, numDays: Int): Int = {
	def toMillisFromJulian(yyddd: Int): Long = {
	def CompressedTimeHHMMSSCC2MilliSecs(compressedTime: Int): Long = {
	def CompressedTimeHHMMSSCC2Secs(compressedTime: Int): Int = {
	def AsCompressedDate(milliSecs: Long): Int = {
	def MonthFromISO8601Int(dt: Int): Int = {
	def YearFromISO8601Int(dt: Int): Int = {
	def DayOfMonthFromISO8601Int(dt: Int): Int = {
	def AgeCalc(yyyymmdd: Int): Int = {
	def iso8601DateFmt(fmtStr : String, yyyymmdds : Int): String = {
	def timestampFmt(fmtStr : String, timestamp : Long): String = {
	def timeStampFromStr(fmtStr : String, timestampStr : String): Long = {
	def dateFromStr(fmtStr : String, timestampStr : String): Long = {
	def timeFromStr(fmtStr : String, timestampStr : String): Long = {
	def dateSecondsSinceMidnight(fmtStr : String, timestampStr : String): Long = {
	def dateSecondsSinceMidnight(timestamp : Long): Long = {
	def millisecsBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def secondsBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def minutesBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def hoursBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def daysBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def weeksBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def monthsBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {
	def yearsBetween(time1 : Long, time2 : Long, inclusive : Boolean = false) : Long = {

timeStampFromStr, dateFromStr, and timeFromStr will parse string representations to Long (millisecs from the epoch)
timestampFmt will print a millisec from epoch Long according to the given format.  Run the scaladoc or browse the
PmmlUdfs.scala file for function semantics.

There is a system type test harness available that tests all functions in trunk/Pmml/src/test/resources.  It is a dropin
type pmml application to be placed beside the other apps in the <installdir>/input.

It is now possible to build your custom udf libraries in separate projects.  The prior requirement to derive from UdfBase
is no longer necessary.  UdfBase has been marked as deprecated and will be eliminated in a future release.  See the object
declaration in PmmlUdfs.scala for example how to change your existing udfs.  The changes to the UdfExtract accomplished this.
A Custom Udf project has been added in the trunk/SampleApplication. It can be used as a template for the new UDF declaration
and also shows the new UDF lib approach.

While testing, bug #357 was fixed... this is the problem using Int, Long, etc in the dataType attribute of the derived fields.

While testing, it was discovered that the easy installer was woefully out of date in terms of the libraries that it needs for the
latest builds.  allDeps.scala and sbtProjDependencies.scala were modified slightly to produce a replacement for the
clusterconfig template's cluster jars list... json style lists with escaped quotes can not be produced.  The list of
libraries to be copied in the easy install script itself has been replaced with the allDeps.scala output generation.  All
project jars and the jars that they depend upon are now in the list as of May 7 2015.

KUNAL's branch that changes some names in the easy scripts and configs should be carefully merged with these changes.

Ahmed's discovery of the null string issue has been addressed with a slight change to the xml event parser in the pmml compiler. The
string length field has been implemented that checks for null and if so returns 0.  Non null strings return actual length of string.

CustomUdfs.scala had some of its functions moved to the core udfs.