Fatafat 48 - remove hierarchical support for msgs
When both a Fixed message or container appears in macro with a collection (e.g., scala.Array, scala.collection._),
the macro logic was choosing the Mapped macro template.  This is because the collections are kinds of ContainerTypeDef which
unfortunately all answer "false" to IsFixed. Rather than change the IsFixed responses in the mdelems, the logic now
checks for "collection-ness" and returns false.  It truly must take a mapped message or mapped container to get the
mapped template now.

What happens when both mapped and fixed appear? The tie goes to mapped. A more thorough type by type treatment of the
arguments are needed.