fix for #394/#452
1) pmmludfTests has new test for the macros with functions.  Currently only macroStateSimple.json/macroTests.xml combination
is functioning.. other tests with the containers in progress.
2) new function/macro signature generation variant added.. promote only member types that are containers, leaving the scalars
alone... this is used for functions like MapKeys.  The intent is to see that the more specific function signature with signature
Map[Int,Any] is chosen over Map[Any,Any] which casts away the types.
3) A number of functions that were defined with both Template and "Any" arguments have been changed.  Templates are now preferred
mechanism.  This does raise the issue that metadata needs to be generated for all the variations of the template that we are
likely to see (e.g., scalar types).  The metadata effectively becomes the template "instantiation" for the Fatafat type system.
Changes to function/function type information generation in the MethodExtractor and ideally in the Message/Container generation
is needed (e.g., we want function "instantiations" for new messages that would cast a list of SomeMessage to an Array of same
message.  There are several work items here that will be filled out and a more serious presentation of the type system for documentation.
4) With this commit, the message and container types have not been addressed.  A follow up commit is forthcoming that addresses
these issues.