Fatafat-118 - properly manage missing value
Overhauled PmmlRuntimeDecls to seek missing value if the variable has not been set.
The mining field case is the focus on this one.  Apply elements in future sprint.
There are baseType defaults if a missingValueReplacement attribute is not available on the mining field
or if there is no mining field at all for the given field in question.
When missingValueReplacement is specified, a suitable DataValue is prepared according to the type
the mining field references.
The dateTime, time, and date "types" are now supported.  They all are aliases of long. Each will be
expressed as millisecs since the 1970 epoch.  String coercion for these time/date based values are
automatically parsed for some of the common formats.  Included are:

    val timeStampPatterns : Array[String] = Array[String]("yyyy-MM-dd HH:mm:ss:SSS"
 							, "yyyy-MM-dd HH:mm:ss"
 							, "MM/dd/yy HH:mm:ss"
 							, "dd-MM-yyyy HH:mm:ss"
 							, "dd-MM-yyyy HH:mm:ss:SSS"
 							, "dd-MMM-yyyy HH:mm:ss"
							, "dd-MM-yyyy HH:mm:ss:SSS"
							)
    val datePatterns : Array[String] = Array[String]("yyyy-MM-dd"
						, "yyyy-MMM-dd"
						, "MM/dd/yy"
						, "MMM/dd/yy"
						, "dd-MM-yyyy"
						, "dd-MMM-yyyy"
						)
    val timePatterns : Array[String] =      Array[String]("HH:mm:ss"
							, "HH:mm:ss:SSS"
							, "h:mm:ss"
							, "h:mm:ss aa"
						)

There is a dearth of Euro date formats. We can tune this.  Even without exact need, the plan
is to allow meta-commands in the missingValueReplacement attribute value that will cause a
DervivedField to execute at runtime that will allow arbitrary format for the date according to
the joda format notation.

Since Fatafat has complex types (messages, containers, collections of any type, etc), another meta-command
is anticipated that will allow one to specify code snippet to initialize such type instances.  For example,
missingValueReplacement="code:Array[Long]()" will cause an AnyDataValue to be created with a value
equal to an empty array of long.

Another test pmml is available that demonstrates the missingValueReplacement capabilities on the mining fields.