NewAPI trait integration into PmmlCompiler
The new traits are now used during generation.  A alternate constructor is introduced that
is the one used by the engine when it calls the CreateNewModel method.  That constructor in turn
calls the primary constructor which has the same arguments as before (unpacked in alt ctor) plus the
two new arguments for the ModelBase abstract class.