partial fix for #529 ... long constant values
Long constants, when values exceed size of Int.Max do not have sufficient type information wrapped around them for
the Scala compiler to recognize it is working with Long, not Int.  This fix changes the macro generation for
long constants such that a long constant will be suffixed with an 'L' when used in an expression.

To do:
1) same thing for expressions in general that are Apply functions.  Ahmed's post illustrates this with the
GreaterThan snippet.