Fatafat 179 Initial revision
Included here is a new SampleApplication directory that contains a jpmml-evalutator
driven model.  This version is fixed in nature, interpreting a TreeModel on the Iris
dataset.

The code builds; the message is prepared; testing today.

Experimenting with the jpmml-evaluator shows that the documentation seems to be a bit ahead of
the implementation found in master.  I could not get the Evaluator factory to determine which
model I was using on its own.  Currently it is explicitly creating a TreeModelEvaluator.

It may be that this capability is in his dev branch which is private.  We could find out.
The other possibility is to write a small parser (steal some code from the PMML compiler) and
traverse the nodes to find the model type.  This would mean supporting at least the
*Model nodes for all of the models... this is maybe a days worth of work.