Fatafat179 Add Java Model for PmmlDecisionTreeModel works
1. Dan helped isolate the scabrous piece of code that was flummoxing
the JavaMirror (the block of code caused a AssertError to be thrown during
introspection).  Commenting it out allows the JPmmlDecisionTreeModel model
to be added to the metadata now.

2. The compile instructions (json) have been updated to include ALL dependencies
for the JPmmlDecisionTreeModel needed for both compile and runtime.

3. The ClusterConfig json was spruced up to include its latest dependencies for
the engine.

4. The metadata api config properties files reflect my build... if you wish
to copy SampleApplication/JpmmlExample/src/main/resources/jpmmlSample
to the <installdir>input/ folder for testing.

5) There is a copy of the JPmmlDecisionTreeModel.java in both the src/main/java/...
and src/main/resources/jpmmlSample/... These should be the same.  I placed the one
in the resources to make it easy to add it to the metadata.

As explained in issue #584, while the compile works, the execution is failing.  We
need runtime dependencies on the model def's dep jars.