Fatafat 131 Udf extraction tool
The tool now generates namespaces using the full package name of the object that
contains the udfs.  The namespace parameter has been eliminated.

 The types generated still use the builtin "System" namespace.

To test, create the types and function json on the CustomUdfs.scala.  One type
will be added (for arrayofinteger) and the functions are all.  A little cursing is
currently required to make this work.