CaseSensitive PmmlCompiler
The pmml compiler will, with this change, find the AttributeDefs for a given FieldRef expression (e.g., msg.DeysnPuf_Id)
and look it up in a case insensitive way.  Whatever is the AttributeDef name is used for the scala generation.

The only issue with this approach is if the black hat tester were to ingest a "case sensitive" message and use field names
with same spelling, differing only in their case usage (e.g., msg.foo and msg.Foo).  In a case sensitive world these are
legal names.  In this version, an exception is thrown if the MessageDef or MappedMsgDef encounters this in their respective
"attributeFor(name:String,caseSensitive:Boolean):AttributeDef" functions. These are two new functions and are called by
the PmmlCompiler fieldref rendering function

By rights, the MessageDef and MappedMsgDef instances should be marked with their respective "caseSensitive" state that was
used at ingestion time.  Then the "attributeFor(name:String,caseSensitive:Boolean):AttributeDef" becomes
"attributeFor(name:String):AttributeDef" where the caseSensitive state used to decide how to match the AttributeDefs in
the message/container is obtained from the message/container instance itself.

This will require a change MappedMsgDef and MessageDef, modification of the MdMgr functions that add/change these, and the
code that uses these Add/Change functions... principally MetadataAPI, MessageDef, and possibly the Bootstrap.