fatafat 224 - kamanja pmml models idempotent
1. New attribute added to derived fields - cacheHint="retain" ... if set to "retain" the value that has been
computed by executing the model for that derived field is preserved.
2. New method in PmmlRuntime's Context class - reset().  Based upon the value of the "retain" state, the value
is reset to "not set" so that on the next use of the instance it will be recomputed.  If "retain" is set, the value
is kept and used as the result for subsequent model instance uses.
3. To test the PMML models in the medical app, the CustomUdfs jar was added to the MetadataAPIConfig properties CLASSPATH.
4. A new test case is provided that illustrates the use - COPDv4.xml

Comments:

By preserving the values of derived values that have been marked "retain", subsequent uses will no longer execute those
so marked... instead using the value computed on the first encounter.  As a result the value will be used for as long as
the model instance is in use... that is, until the model has been replaced/removed with a newer version or the cluster is
brought down.

In the test case (COPDv4.xml), the "dimensional" filter values pulled into the model for filtering copd symptoms are marked
with cacheHint="retain" attributes.  By doing this, the costly marshalling/demarshalling the filter values from EnvContext is
avoided.  The filter sets, after the first execution, are already initialized.

This begs the question:  "What if I want to periodically change the values of my filter sets or other 'expensive' data that I
made part of my cached instance?"  This is exactly the situation illustrated with twitter demo's need to peridically change topics
and search criteria for those topics based upon review of preceding results/new business requirements.

There are likely a number of ways to do this, but the most obvious and least intrusive to our current mechanisms is to allow
models to accept more than one message.  In other words, the model processes the conventional data message (e.g., a tweet) as well as
a control message that tells the model to pick up new filter or search data or other bits of information that are important for controlling
the model.

A minor metadata change is needed and the possibility that more than one message can be "handled" by any given model.