Fix for #781
Object generator in the pmml node printer helpers was not accessing package name variable established in
the model comment function.

Metadata api template fixed so that the copd v3 model could be compiled.
The copd v3 model changed to use an explicit namespace in the application name of the header.

This header:

	<Header copyright="Medco Corp. Copyright 2014" description="COPD Risk Assessment">
		<Application name="com.medco.pmml.COPDRiskAssessmentNmSpc" version="00.01.00"/>
	</Header>

now produces this in the Object:

object COPDRiskAssessmentNmSpc extends ModelBaseObj {
    override def ModelName(): String = "com.medco.pmml.COPDRiskAssessmentNmSpc"
    override def Version(): String = "1000000"
   .
   .
   .